Refactor: auth 관련 로그인, response dto 이동(auth/member/dto -> auth/dto)
