+++ src/main/java/com/chs/cafeapp/auth/controller/AuthAdminController.java
@@ -3,9 +3,8 @@ package com.chs.cafeapp.auth.controller;
 import com.chs.cafeapp.auth.dto.LogOutResponse;
 import com.chs.cafeapp.auth.dto.PasswordEditInput;
 import com.chs.cafeapp.auth.dto.PasswordEditResponse;
-import com.chs.cafeapp.auth.member.dto.AuthResponseDto;
-import com.chs.cafeapp.auth.member.dto.SignInRequestDto;
-import com.chs.cafeapp.auth.service.AuthService;
+import com.chs.cafeapp.auth.dto.AuthResponseDto;
+import com.chs.cafeapp.auth.dto.SignInRequestDto;
 import com.chs.cafeapp.auth.service.AuthTokenService;
 import com.chs.cafeapp.auth.service.impl.AuthAdminService;
 import com.chs.cafeapp.auth.token.dto.TokenResponseDto;
