+++ src/main/java/com/chs/cafeapp/auth/controller/AuthMemberController.java
@@ -6,8 +6,8 @@ import com.chs.cafeapp.auth.service.impl.AuthMemberService;
 import com.chs.cafeapp.exception.CustomException;
 import com.chs.cafeapp.auth.token.dto.TokenResponseDto;
 import com.chs.cafeapp.auth.member.dto.SignUpRequestDto;
-import com.chs.cafeapp.auth.member.dto.SignInRequestDto;
-import com.chs.cafeapp.auth.member.dto.AuthResponseDto;
+import com.chs.cafeapp.auth.dto.SignInRequestDto;
+import com.chs.cafeapp.auth.dto.AuthResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
