+++ src/main/java/com/chs/cafeapp/auth/service/impl/AuthAdminService.java
@@ -18,8 +18,8 @@ import com.chs.cafeapp.auth.component.TokenBlackList;
 import com.chs.cafeapp.auth.component.TokenPrepareList;
 import com.chs.cafeapp.auth.dto.PasswordEditInput;
 import com.chs.cafeapp.auth.dto.PasswordEditResponse;
-import com.chs.cafeapp.auth.member.dto.AuthResponseDto;
-import com.chs.cafeapp.auth.member.dto.SignInRequestDto;
+import com.chs.cafeapp.auth.dto.AuthResponseDto;
+import com.chs.cafeapp.auth.dto.SignInRequestDto;
 import com.chs.cafeapp.auth.member.dto.SignUpRequestDto;
 import com.chs.cafeapp.auth.service.AuthService;
 import com.chs.cafeapp.auth.token.dto.TokenDto;
@@ -30,15 +30,12 @@ import com.chs.cafeapp.exception.CustomException;
 import com.chs.cafeapp.security.TokenProvider;
 import java.time.LocalDateTime;
 import java.util.Collection;
-import java.util.Collections;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
