+++ src/main/java/com/chs/cafeapp/auth/service/impl/AuthMemberService.java
@@ -26,8 +26,8 @@ import com.chs.cafeapp.auth.component.TokenBlackList;
 import com.chs.cafeapp.auth.component.TokenPrepareList;
 import com.chs.cafeapp.auth.dto.PasswordEditInput;
 import com.chs.cafeapp.auth.dto.PasswordEditResponse;
-import com.chs.cafeapp.auth.member.dto.AuthResponseDto;
-import com.chs.cafeapp.auth.member.dto.SignInRequestDto;
+import com.chs.cafeapp.auth.dto.AuthResponseDto;
+import com.chs.cafeapp.auth.dto.SignInRequestDto;
 import com.chs.cafeapp.auth.member.dto.SignUpRequestDto;
 import com.chs.cafeapp.auth.member.entity.Member;
 import com.chs.cafeapp.auth.member.repository.MemberRepository;
